backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-core==0.3.15
langgraph==0.2.45
pydantic==2.10.2
requests==2.32.3
httpx==0.27.2
